apps/api/cdk/requirements.txt
# AWS CDK requirements for LCopilot Monitoring Stack
aws-cdk-lib>=2.100.0
constructs>=10.0.0
boto3>=1.26.0